cooffeh/requirements.txt
jurigged
Pyrogram
python-dotenv
pytube
beautifulsoup4
requests
Pillow
qrcode
scrapingant-client
pyshorteners